Everything should be working and all clean code
Have cleaned up the comments in the code and all the different types of
facts should be functional and updateable.  This should also support
`puppet resource fact`.  Supported fact types all go to YAML right now
and supported fact types are as such: String, Array, Hash and mixtures
of Arrays and Hashes.